Address race conditions with edge allocations:
1. Locking for conflicting IP's with subnet creation
2. Router and DHCP share the same pool. The lock for those is now
   combined
3. Locking DHCP adges for updates

+++ vmware_nsx/neutron/plugins/vmware/dbexts/nsxv_db.py
@@ -74,8 +74,9 @@ def update_nsxv_router_binding(session, router_id, **kwargs):
 def delete_nsxv_router_binding(session, router_id):
     with session.begin(subtransactions=True):
         binding = (session.query(nsxv_models.NsxvRouterBinding).
-                   filter_by(router_id=router_id).one())
-        session.delete(binding)
+                   filter_by(router_id=router_id).first())
+        if binding:
+            session.delete(binding)
 
 
 def get_edge_vnic_binding(session, edge_id, network_id):
